MQTT: Reinstate 'total_increasing' support for kWh sensors
This was reverted due to a ZWave-JS-UI bug where it advertised sensors
with 'total_increasing' even when they shouldn't be. So this commit
reinstates it except *not* for ZWave devices.

A state class of "total_increasing" indicates "a monotonically increasing
increasing positive total which periodically restarts counting from 0" …
"A decreasing value is interpreted as the start of a new meter cycle or the
replacement of the meter".

Handle this by storing an "epoch" for the measurement; this is the base
value added to the raw measurement, to reach the actual total to be
recorded.

If the result of adding a measurement to this epoch is ever *lower* than
the previous recorded value, that means the counter must have restarted,
so the epoch is increased to match the previous value and we keep
counting from there.

The epoch is stored in "StrParam1" of the DeviceStatus record.

Clean up the matching of power and usage sensors to use the *actual*
value of the paired sensor instead of its raw last_value. This keeps
knowledge about one type of sensor away from the other type. So the
code handling the Watt sensor no longer needs to care about different
multipliers for kWh/Wh/Wm sensors, etc. — and, more to the point, doesn't
need to learn about the epoch handling. It was already failing to throw
away bogus (< -1000000) readings as the paired sensor should.

Fixes: #6160