Fix MQTT AutoDiscover device vs. sensor name confusion
During MQTT AutoDiscovery, we were overwriting the name of the parent MQTT
device with the name of each sensor we discovered that was attached to it.

Each time we discovered a sensor, we get the device information again in
that sensor's JSON information, and construct a name using what we find
in the JSON, e.g. "device (sensor_name)".

But then we put that sensor-specific name into pDevice->name which we use
to represent the *device*, not just that sensor.

This mostly didn't matter, as pDevice->name is *mostly* a write-only
string, because sensor naming just uses the device name from the JSON
each time.

The only case where we care about pDevice->name is when we build up
combined Temperature/Humidity/Barometer devices from multiple MQTT
sensors. In *that* case, we were constructing the name of the combined
Temp+Hum+Baro device from pDevice->name, which contained the name of
the last *sensor* that we happened to find attached to that device!

Which explains why we ended up with Temp+Hum+Baro devices with weird
names like "bathroom (Blue LED)".

Clean up the device naming, use a local variable for sensor_name instead
of messing with pDevice->name, and make it a bit easier to follow and
better documented.

Fixes: #6164